dev-requirements.txt
numpy
pyserial-asyncio
pytest
pytest-asyncio
mypy
black
ruff
